requirements.txt
Pillow
Django
django-livesync
pydevd
django-crispy-forms
